oj_backend/requirements.txt
Django>=5.2.4
google-generativeai>=0.8.5
psutil>=5.9.0
python-dotenv>=1.0.0 